rename formatNumberToLocale to formatNumberForLocale
