+++ lib/Core/TableColumn.js
@@ -16,7 +16,7 @@ var knockout = require('terriajs-cesium/Source/ThirdParty/knockout');
 var TimeInterval = require('terriajs-cesium/Source/Core/TimeInterval');
 var TimeIntervalCollection = require('terriajs-cesium/Source/Core/TimeIntervalCollection');
 
-var formatNumberToLocale = require('../Core/formatNumberToLocale');
+var formatNumberForLocale = require('../Core/formatNumberForLocale');
 var getUniqueValues = require('../Core/getUniqueValues');
 var inherit = require('../Core/inherit');
 var VarType = require('../Map/VarType');
@@ -695,7 +695,7 @@ function getFormattedValues(tableColumn) {
             if (isNaN(value)) {
                 return value;
             }
-            return formatNumberToLocale(value, tableColumn.format);
+            return formatNumberForLocale(value, tableColumn.format);
         });
     } else if (tableColumn.type === VarType.TIME) {
         return tableColumn.dates.map(function(date, index) {
