+++ lib/Core/formatNumberForLocale.js
@@ -18,7 +18,7 @@ function toLocaleStringSupportsOptions() {
 *        https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString .
 *        Note we default useGrouping to false (not true) and maximumFractionDigits to 20 (not 3).
 */
-function formatNumberToLocale(number, options) {
+function formatNumberForLocale(number, options) {
     if (number === null) {
         return '';
     }
@@ -54,4 +54,4 @@ function formatNumberToLocale(number, options) {
     return str.replace(/\B(?=(\d{3})+(?!\d))/g, ',') + frac;
 }
 
-module.exports = formatNumberToLocale;
+module.exports = formatNumberForLocale;
