+++ lib/Core/formatPropertyValue.js
@@ -2,18 +2,18 @@
 
 /*global require*/
 var linkifyContent = require('./linkifyContent');
-var formatNumberToLocale = require('./formatNumberToLocale');
+var formatNumberForLocale = require('./formatNumberForLocale');
 
 /**
 * Format the value for the description, used by the Feature Info Panel.
 * Strings have markdown applied to them. Anything else is returned as-is.
 *
 * @param {} value The value to format.
-* @param {Object} [options] Number formatting options, passed to formatNumberToLocale.
+* @param {Object} [options] Number formatting options, passed to formatNumberForLocale.
 */
 function formatPropertyValue(value, options) {
     if (typeof value === 'number') {
-        return formatNumberToLocale(value, options);
+        return formatNumberForLocale(value, options);
     } else
     if (typeof value === 'string') {
         // do not linkify if it contains html elements, which we detect by looking for <x...>
