+++ lib/Map/LegendHelper.js
@@ -8,7 +8,7 @@ var defined = require('terriajs-cesium/Source/Core/defined');
 var definedNotNull = require('terriajs-cesium/Source/Core/definedNotNull');
 
 var ColorMap = require('../Map/ColorMap');
-var formatNumberToLocale = require('../Core/formatNumberToLocale');
+var formatNumberForLocale = require('../Core/formatNumberForLocale');
 var Legend = require('../Map/Legend');
 var TableStyle = require('../Models/TableStyle');
 var VarType = require('../Map/VarType');
@@ -337,7 +337,7 @@ function convertToStringWithAtMostTwoDecimalPlaces(f, tableColumnStyle) {
     } else {
         options = {maximumFractionDigits: 2};
     }
-    return formatNumberToLocale(f, options);
+    return formatNumberForLocale(f, options);
 }
 
 function convertColorArrayToCssString(colorArray) {
